Работает по номеру смены, а должно по номеру сотрудника, обкатал ошибки

+++ log-entry-service/src/main/java/com/example/logentryservice/model/LogEntry.java
@@ -30,7 +30,7 @@ public class LogEntry {
     private LocalDateTime endTime;
 
     @Column(name = "employee_id", nullable = false)
-    private int employeeId;
+    private Long employeeId;
 
     @Column(name = "message")
     private String message;
